+++ guillotina/db/storages/pg.py
@@ -1,6 +1,8 @@
+from async_timeout import timeout
 from asyncio import shield
 from guillotina._settings import app_settings
 from guillotina.const import TRASHED_ID
+from guillotina.db.events import StorageCreatedEvent
 from guillotina.db.interfaces import IPostgresStorage
 from guillotina.db.storages.base import BaseStorage
 from guillotina.db.storages.utils import clear_table_name
@@ -8,6 +10,7 @@ from guillotina.db.storages.utils import get_table_definition
 from guillotina.db.storages.utils import register_sql
 from guillotina.db.storages.utils import SQLStatements
 from guillotina.db.uid import MAX_UID_LENGTH
+from guillotina.event import notify
 from guillotina.exceptions import ConflictError
 from guillotina.exceptions import ConflictIdOnContainer
 from guillotina.exceptions import TIDConflictError
@@ -59,7 +62,7 @@ WHERE of = $1::varchar({MAX_UID_LENGTH})
 register_sql(
     "GET_CHILD",
     f"""
-SELECT zoid, tid, state_size, resource, type, state, id
+SELECT zoid, tid, state_size, resource, type, state, id, parent_id, of
 FROM {{table_name}}
 WHERE parent_id = $1::varchar({MAX_UID_LENGTH}) AND id = $2::text
 """,
@@ -68,7 +71,7 @@ WHERE parent_id = $1::varchar({MAX_UID_LENGTH}) AND id = $2::text
 register_sql(
     "GET_CHILDREN_BATCH",
     f"""
-SELECT zoid, tid, state_size, resource, type, state, id
+SELECT zoid, tid, state_size, resource, type, state, id, parent_id, of
 FROM {{table_name}}
 WHERE parent_id = $1::varchar({MAX_UID_LENGTH}) AND id = ANY($2)
 """,
@@ -97,7 +100,7 @@ WHERE zoid = $1::varchar({MAX_UID_LENGTH})
 register_sql(
     "GET_ANNOTATION",
     f"""
-SELECT zoid, tid, state_size, resource, type, state, id, parent_id
+SELECT zoid, tid, state_size, resource, type, state, id, parent_id, of
 FROM {{table_name}}
 WHERE
     of = $1::varchar({MAX_UID_LENGTH}) AND
@@ -737,37 +740,8 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
                 self._supports_unique_constraints = True
                 await self.initialize_tid_statements()
                 await self.read_conn.execute(trash_sql)
+                await notify(StorageCreatedEvent(self))
 
-            # migrate to larger VARCHAR size...
-            result = await self.read_conn.fetch(
-                """
-    select * from information_schema.columns
-    where table_name='{}'""".format(
-                    self._objects_table_name
-                )
-            )
-            if len(result) > 0 and result[0]["character_maximum_length"] != MAX_UID_LENGTH:
-                log.warn("Migrating VARCHAR key length")
-                await self.read_conn.execute(
-                    f"""
-    ALTER TABLE {self._objects_table_name} ALTER COLUMN zoid TYPE varchar({MAX_UID_LENGTH})"""
-                )
-                await self.read_conn.execute(
-                    f"""
-    ALTER TABLE {self._objects_table_name} ALTER COLUMN of TYPE varchar({MAX_UID_LENGTH})"""
-                )
-                await self.read_conn.execute(
-                    f"""
-    ALTER TABLE {self._objects_table_name} ALTER COLUMN parent_id TYPE varchar({MAX_UID_LENGTH})"""
-                )
-                await self.read_conn.execute(
-                    f"""
-    ALTER TABLE {self._blobs_table_name} ALTER COLUMN bid TYPE varchar({MAX_UID_LENGTH})"""
-                )
-                await self.read_conn.execute(
-                    f"""
-    ALTER TABLE {self._blobs_table_name} ALTER COLUMN zoid TYPE varchar({MAX_UID_LENGTH})"""
-                )
             self._connection_initialized_on = time.time()
 
     async def initialize_tid_statements(self):
@@ -782,21 +756,17 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
 
     @restart_conn_on_exception
     async def open(self):
-        conn = await self.pool.acquire(timeout=self._conn_acquire_timeout)
-        return conn
+        async with timeout(self._conn_acquire_timeout):
+            return await self.pool.acquire()
 
     async def close(self, con):
         try:
-            await shield(asyncio.wait_for(self.pool.release(con, timeout=1), 1))
-        except (
-            asyncio.CancelledError,
-            RuntimeError,
-            asyncio.TimeoutError,
-            asyncpg.exceptions.ConnectionDoesNotExistError,
-        ):
-            pass
+            await shield(self.pool.release(con, timeout=1))
+        except (asyncio.CancelledError, asyncio.TimeoutError, asyncpg.exceptions.ConnectionDoesNotExistError):
+            log.warning("Exception on connection close", exc_info=True)
 
     async def terminate(self, conn):
+        log.warning(f"Terminate connection {conn}", exc_info=True)
         conn.terminate()
 
     async def load(self, txn, oid):