+++ guillotina/tests/test_api.py
@@ -155,8 +155,7 @@ async def test_create_content(container_requester):
         assert status == 201
         _, status = await requester("GET", "/db/guillotina/item1")
         assert _["@static_behaviors"][0] == "guillotina.behaviors.dublincore.IDublinCore"
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("item1")
@@ -203,8 +202,7 @@ async def test_put_content(container_requester):
         assert resp[IDublinCore.__identifier__]["tags"] is None
         assert resp[IAttachment.__identifier__]["file"] is None
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("item1")
@@ -311,8 +309,7 @@ async def test_create_contenttype_with_date(container_requester):
             ),
         )
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("item1")
@@ -995,8 +992,7 @@ async def test_field_values_dict_bucket_preconditions(container_requester):
         assert status == 412
 
         # delete annotation for bucket that should be there, we should get 410
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db):
             container = await root.async_get("guillotina")
             obj = await container.async_get("item1")