+++ guillotina/tests/test_attachment.py
@@ -105,8 +105,7 @@ async def test_multi_upload(manager_type, redis_container, container_requester):
         assert status == 200
         assert response == b"Y" * 1024 * 1024 * 10
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("foobar")
@@ -139,8 +138,7 @@ async def test_large_upload_chunks(manager_type, redis_container, container_requ
         assert status == 200
         assert len(response) == (1024 * 1024 * 10)
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("foobar")
@@ -194,8 +192,7 @@ async def test_tus(manager_type, redis_container, container_requester):
         assert status == 200
         assert len(response) == (1024 * 1024 * 10)
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("foobar")
@@ -251,8 +248,7 @@ async def test_tus_multi(manager_type, redis_container, container_requester):
         assert status == 200
         assert len(response) == (1024 * 1024 * 10)
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("foobar")
@@ -317,8 +313,7 @@ async def test_tus_unknown_size(manager_type, redis_container, container_request
         assert status == 200
         assert len(response) == offset
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("foobar")
@@ -449,8 +444,7 @@ async def test_tus_with_empty_file(manager_type, redis_container, container_requ
         assert status == 200
         assert len(response) == 0
 
-        request = utils.get_mocked_request(db=requester.db)
-        root = await utils.get_root(request)
+        root = await utils.get_root(db=requester.db)
         async with transaction(db=requester.db, abort_when_done=True):
             container = await root.async_get("guillotina")
             obj = await container.async_get("foobar")