+++ guillotina/traversal.py
@@ -283,6 +283,7 @@ class MatchInfo(BaseMatchInfo):
                 view_result = await self.view()
                 await commit(warn=False)
             except (ConflictError, TIDConflictError):
+                await abort()
                 # bubble this error up
                 raise
             except response.Response as exc: